rename auto schema file to something more obvious
