+++ coffea/nanoevents/schemas/auto.py
